# backend/outputs/ApiContract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PIA Developer Console API",
    "description": "PIA Platform Observability System \u2014 All data served through stable projections and versioned contracts.",
    "version": "2.0.0",
    "x-stability": "frozen"
  },
  "paths": {
...
    "/api/v1/sync/start": {
      "post": {
        "tags": [
          "Sync Engine"
        ],
        "summary": "Start Sync",
        "description": "Start a sync job. Returns job_id immediately; poll /sync/job/{job_id} for status.",
        "operationId": "start_sync_api_v1_sync_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/job/{job_id}": {
      "get": {
        "tags": [
          "Sync Engine"
        ],
        "summary": "Get Job Status",
        "description": "Sync Explorer: full job status with progress counters.",
        "operationId": "get_job_status_api_v1_sync_job__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncJobDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/cancel/{job_id}": {
      "post": {
        "tags": [
          "Sync Engine"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_job_api_v1_sync_cancel__job_id__post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sync/history": {
      "get": {
        "tags": [
          "Sync Engine"
        ],
        "summary": "Get Sync History",
        "description": "Completed sync jobs.",
        "operationId": "get_sync_history_api_v1_sync_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sync/rate-limit": {
      "get": {
        "tags": [
          "Sync Engine"
        ],
        "summary": "Get Rate Limit",
        "description": "Current GitHub API rate limit status.",
        "operationId": "get_rate_limit_api_v1_sync_rate_limit_get",
        "parameters": [
          {
            "name": "github_token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Github Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "CancelResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "repository": {
            "type": "string",
            "title": "Repository"
          },
          "sync_mode": {
            "type": "string",
            "title": "Sync Mode"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "repository",
          "sync_mode",
          "status",
          "started_at"
        ],
        "title": "JobResponse"
      },
...
      "RateLimitResponse": {
        "properties": {
          "remaining": {
            "type": "integer",
            "title": "Remaining"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "reset_at": {
            "type": "integer",
            "title": "Reset At"
          }
        },
        "type": "object",
        "required": [
          "remaining",
          "limit",
          "reset_at"
        ],
        "title": "RateLimitResponse"
      },
...
      "SyncHistoryResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/SyncJobDTO"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "jobs"
        ],
        "title": "SyncHistoryResponse"
      },
      "SyncJobDTO": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "repository": {
            "type": "string",
            "title": "Repository"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          },
          "sync_mode": {
            "type": "string",
            "title": "Sync Mode"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "type": "string",
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "progress": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Progress"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "repository",
          "workspace_id",
          "sync_mode",
          "status",
          "started_at",
          "completed_at",
          "error",
          "progress"
        ],
        "title": "SyncJobDTO"
      },
      "SyncRequest": {
        "properties": {
          "repository": {
            "type": "string",
            "title": "Repository"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "incremental"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "default": "main"
          },
          "commit_limit": {
            "type": "integer",
            "title": "Commit Limit",
            "default": 100
          },
          "github_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Token"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "type": "object",
        "required": [
          "repository"
        ],
        "title": "SyncRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}